Store ObjectInteraction data as a component on terrain tiles
