+++ emergence_lib/src/graphics/mod.rs
@@ -2,7 +2,7 @@
 
 use bevy::prelude::*;
 
-use crate::{asset_management::AssetState, player_interaction::InteractionSystem};
+use crate::asset_management::AssetState;
 
 use self::lighting::LightingPlugin;
 
@@ -20,8 +20,7 @@ impl Plugin for GraphicsPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(LightingPlugin)
             .add_system(units::display_held_item.run_if(in_state(AssetState::Ready)))
-            .add_system(inherit_materials.in_base_set(CoreSet::PostUpdate))
-            .add_system(selection::display_tile_interactions.after(InteractionSystem::SelectTiles));
+            .add_system(inherit_materials.in_base_set(CoreSet::PostUpdate));
     }
 }
